+++ graph/src/constants.ts
@@ -1,0 +1,3 @@
+export const CONDITIONAL_ITER_LIMIT = 10;
+export const NONCONDITIONAL_PATH_LIMIT = 10;
+export const CONDITIONAL_CLEANUP_ITER_LIMIT = 10;
